libfs_mgr: Fix of an incorrectly opened ZRAM device.
The "max_comp_streams" parameter in the "fstab" file is not taken into account when forming a zram file, since the file is open for reading(re). It is necessary to open the file for writing (re+).
Fixing this bug improves performance on devices with zram support in the kernel.

Change-Id: Idcd45a2b019f80393c83a12d647a1c00dcd56dca
Signed-off-by: SirRGB <sirrgb@proton.me>